21 solved (w help)
my solution in day21.py works but is slow and takes a long time to get the solution to part 2. So I double checked my answers with the code I did not write in day21_works.py

This one took awhile so I do not have energy to optimize further lol.